pyproject.toml
[project]
name = "supervisor-webui"
version = "1.0"
description = "SupervisorWebUI webserver"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9",
    "argparse",
    "requests",
    "paramiko",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.27.1",
    "python-multipart"
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]
extend-exclude = ["venv", ".venv", "build", "dist", "deploy"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PL"]
ignore = ["E203", "B008"]

[tool.ruff.format]
quote-style = "double"